Add quirks for H610A
At the time of writing, the platform API doesn't return correct
information for this device, and refuses to control it appropriately.

This commit adds the concept of a quirks database that can be used
to override the metadata and try to arrive at a better outcome.

This is sufficient to regain color and brightness control over this
device.

For scenes, this commit attempts to use the fruits of some reverse
engineering to compute a scene code to send via an undocumented
part of the LAN API.  In practice, this is most reliable with
scene codes that fit in a single byte; the others likely need
to get routed via the IoT MQTT connection.

That can be examined in a follow up commit.

refs: https://github.com/wez/govee2mqtt/issues/7