explicitly manipulate the addon version number
I'd assumed that addons were versioned just like docker,
but they aren't.  You won't get an option to update if
it is just using "latest" all the time.

I suppose that makes sense for explicit point in time
releases, but it is a bit frustrating for continuous
delivery.

This commit adds a script to make it convenient to
explicitly tag and bump a version number.