Make more robust around reporting scenes
If we fail to figure out what scenes are available, that shouldn't
block bringing up the rest of the stuff that we know.

Improve error context and reporting along this code path.

Make it clear that if we fail up through the mqtt loop that we
are fatally broken, and terminate the process.

refs: https://github.com/wez/govee2mqtt/issues/4