Add segment color and brightness control
Since Govee doesn't report the status of individual segments,
the entities are registered with home assistant in optimistic
mode, meaning that what you last set via home assistant is what
it believes to be the state of the device.

This is likely sufficient for eg: capturing a scene in home assistant,
but you should not use the state of one of the segment lights to
make any kind of automation decisions.