Round out H7160 humidifier support
This seems to get things in a reasonably complete state for me:

* Night light power, color and brightness
* Target humidity percent shown in humidifier entity
* Manual mist level control

Custom is not mapped as an entity; it seems rather complex in
the UI and the metadata is no use in understand how to encode
those behaviors.

This might happen to enable other humidifiers as well,
but I wouldn't be surprised if they need additional care
and attention to work reasonably.