Downgrade platform API trust when it conflicts with LAN API
If the Platform API says that the device isn't an rgb light,
but we have available LAN API, then we decide not to trust
the platform API for that device.

Again, I don't think this will change the outcome in
the linked issue because this primarily affects us
when trying to change scenes.

refs: https://github.com/wez/govee2mqtt/issues/32