Add read-after-write capability for platform api
There are a number of devices for which we don't know how to
fully decode data from the IoT API, but where the Platform API
will return us the data.

Contrive to read that data both in the periodic polls and also
to read it back after we've written a control request to the
device, so that we can more immediately reflect the new state.