platform api now errors for my humidifier
It now yields:

```
{"requestId":"uuid","msg":"devices not support this instance","code":400,
```

which causes the previously cached-for-me empty results to be returned,
which is fine, but since this request failed it uses the negative ttl
of 1 minute, so we make this request every time we poll devices,
which is bad from an api quota limit perspective.

I could change the logic in the methods that fetch scenes to check
for this code and message string, but that seems brittle and error
prone.

So this commit takes a dual-pronged approach:

1. Bump the negative ttl to 5 minutes for any error in fetching scenes
2. Add in a quirk capability so that we know whether a device
   definitively does or does not support scenes and populate that
   for this particular device.  If we know it doesn't support them,
   we simply skip trying to populate scene metadata.