fix "devices not support this instance" error message
This was actually harmless, but noisy.  The cause was that Govee
API returns this error message when trying to list the available
scenes for the device. When I started this project the API returned
an empty list, but as of a few days ago they made it return an
error condition for devices that don't support scenes. We ignore
the error and continue, so it didn't harm anything, it was
just noisy.

refs: https://github.com/wez/govee2mqtt/issues/58
refs: https://github.com/wez/govee2mqtt/issues/50
refs: https://github.com/wez/govee2mqtt/issues/59
refs: https://github.com/wez/govee2mqtt/issues/56