switch last-will to unretained
There are a couple of users who report that their devices appear
as offline in hass seemingly for no good reason.

As far as I can see from the logs, they both appear to otherwise
be in good working order.

One of them has troubles that correlate with restarting hass
itself and/or the broker addon.  It seems like there may not
be good connection recovery in that situation.

The other seems to have no issues at all, except that the availability
topic somehow reverts back to offline.

What I suspect is some kind of intermittent network burp that somehow
triggers the last will message to be sent by the broker.  I'm also
a bit dubious of the retained nature of that message: I've read a
few things online where folks are working around retained messages
that somehow end up in a sticky state.

So, this commit turns off the retained flag in the hope that we
don't see any additional weird reports like this going forwards.

refs: https://github.com/wez/govee2mqtt/issues/37
refs: https://github.com/wez/govee2mqtt/issues/60