Potentially fix temperature sensor scaling
This commit introduces temperature (and humidity) units awareness
to the quirks layer, and defines some quirks for a handful of
devices.

These quirks allow us to correctly interpret the sensor readings
returned from the platform API for these devices, which do not
have any other metadata available from Govee to aid with this.

refs: https://github.com/wez/govee2mqtt/issues/58
refs: https://github.com/wez/govee2mqtt/issues/56
refs: https://github.com/wez/govee2mqtt/issues/47

It may also help to avoid misinterpreting the device power state
with the rgb light state on devices that are not primarily
lights, such as certain kinds of space heaters and humidifiers,
and with some light strips.

refs: https://github.com/wez/govee2mqtt/issues/59
refs: https://github.com/wez/govee2mqtt/issues/49