parallelize initial lan disco device querying
I noticed that sometimes one or two of my devices would show
up with a warning about not responding on the lan api yet.
I think this is a consequence of a couple devices being a
bit tardy, and since we were polling them sequentially, we
could run out of time (we allow 10 seconds before printing
the initial status) to hear from them.

This commit runs that polling in parallel.  It's possible
that this will need to be tweaked in the future to apply
a semaphore to limit just how many devices we're trying
to reach at once; for homes with a large number of govee
devices on the lan this might get a bit gnarly.

We'll see!