make iot mqtt subscriber more robust
I was trying to improve debugging for reverse engineering purposes
and noticed that we were silently breaking.

What was happening was that the server was disconnecting us after
attempting to subscribe to device topic(s), and then the mqtt
client would leave our original subscription request hanging
indefinitely.

I've upgraded the mqtt client to cancel all pending waiters
on disconnect, and moved things around in here so that we can
catch and log this kind of error more assertively.