improve platform api support for Humidifiers
Where we don't have working IoT packet parsers for a humidifier,
we can fall back to the platform API.

Based on the capabilities reported for my device, that allows us
to report Manual mister level and mode, and to set (but not report)
the target humidity.

The write-only aspect of that API is a bit frustrating, but we
can set the entity up in optimistic mode and it seems to work
out OK in the hass UI.

Adjust the platform device polling to delay a few seconds before
polling; for my humidifier it takes a little bit for the API
to report the changed power state for my device if it was activated
as a side effect of setting the target humidity.

refs: https://github.com/wez/govee2mqtt/issues/22
refs: https://github.com/wez/govee2mqtt/issues/49 (maybe)