Fix potential integer underflow with target humidity
I think that the 128+value thing is the MSB being set rather than
being an offset value.

I saw an underflow happen once during testing, but it didn't emit
a stack trace and has happened again since, which suggests to me
that the device may not always set that bit.