Fix "has no codec for TypeId" errors
We can resolve these by catching and ignoring the error from the encode
attempt; the only reason it can fail is if we don't know how to encode
for a given device.

The we fallback to whatever other strategy is available: most likely
it will be a platform API call.

refs: https://github.com/wez/govee2mqtt/issues/73
refs: https://github.com/wez/govee2mqtt/issues/56