speculatively support changing target temperature
For devices that report temperature settings, this commit
will now expose a number entity in Celsius that will change
the target temperature when set.

None of the govee device metadata I've see so far report
the current temperature target, so this entity works in
optimistic mode.

It's possible that hass will keep it greyed out until
a value is reported for it, which will be kinda sucky.

refs: https://github.com/wez/govee2mqtt/issues/30
refs: https://github.com/wez/govee2mqtt/issues/47
refs: https://github.com/wez/govee2mqtt/issues/56
refs: https://github.com/wez/govee2mqtt/issues/59
refs: https://github.com/wez/govee2mqtt/issues/69
refs: https://github.com/wez/govee2mqtt/issues/75