Add Temperature Scale option
* *Important*: when upgrading from an earlier version,
  you should delete any devices that have temperature
  sensors or controls and then stop the addon/docker container
  and ensure that you have set the temperature scale to your
  preference, then start the addon/container again.

This commit introduces a configuration option to specify your
preferred temperature scale as either "C" or "F" for Celsius
or Farenheit, respectively.

Entities created in home assistant will be set to use those units,
and convert to the underlying device units.

Since I don't own any devices with this characteristics, this change
has not been tested with real devices and it may not work; please
file an issue if you're having trouble with this!

refs: https://github.com/wez/govee2mqtt/issues/30
refs: https://github.com/wez/govee2mqtt/issues/47
refs: https://github.com/wez/govee2mqtt/issues/56
refs: https://github.com/wez/govee2mqtt/issues/59
refs: https://github.com/wez/govee2mqtt/issues/69
refs: https://github.com/wez/govee2mqtt/issues/75