Add more delays on startup
There's been a number of reports of things showing as unavailable
after a haos reboot, let's try to slow things down a bit in the
interest of hass more reliably accepting the registration.

It's frustrating because the various mqtt events are supposed
to be the signal that things are ready, and they aren't!