improve undoc and platform API errors on startup
This commit expands upon the error message reported during failed
initial disco via those APIs to explain why we cannot automatically
continue with the service startup, and to suggest the workaround, while
pointing at the tracking issue.

I'd wanted to make this non-blocking (you can see evidence of support
for that in this commit), but then I re-discovered the tracking issue
and the impact that such a thing would have on entity ids.  I don't
think it is a good idea to automatically proceed with altered entity
ids as it could wreak havoc with automations.

We now also will periodically (every 10 minutes) query those APIs for an
updated list of devices/rooms, which helps to detect newly adopted govee
devices after the service has started.

refs: https://github.com/wez/govee2mqtt/issues/76