addon: try to handle weird startup timing issue better
There are a handful of reports of:

  "Error: Please specify the mqtt broker"

that happen when using the default implicit broker credentials
as described in the docs.

Some users report that this issue occurs on startup of haos
and that restarting or deferring startup of govee2mqtt will
resolve this.

Other users report that explicitly configuring credentials
and the mqtt endpoint resolves things for them.  I suspect
this latter case is actually a separate issue, but that it
manifests in a very similar way.

Looking at zigbee2mqtt I can see that it enables the hassio api option
in its addon configuration; it's really unclear from the addon
development guide docs when this is needed because for me and most other
users, the addon works fine without it.  I've set it to true in this
commit as a speculative measure.

The other thing that I saw when reviewing zigbee2mqtt was this issue:
https://github.com/home-assistant/supervisor/issues/3884 where it
appears as though there is some kind of startup race condition between
addons, despite the addon configuration stating that mqtt is required
for us to start.  Their resolution was to use the bashio library to
assert that a basic config option was present.  We can't do literally
that, but we can check that either the mqtt broker is up or that the
mqtt_host option is set, and bail out early if both conditions are not
met.

Hopefully that will persuades HAOS's supervisor watchdog to try us again
in a little bit, and hopefully make the mqtt details available to us.

refs: https://github.com/wez/govee2mqtt/issues/496
refs: https://github.com/wez/govee2mqtt/issues/486
refs: https://github.com/wez/govee2mqtt/issues/471
refs: https://github.com/wez/govee2mqtt/issues/464
refs: https://github.com/wez/govee2mqtt/issues/408
refs: https://github.com/wez/govee2mqtt/issues/142