Revise error message for missing broker
I just spent most of the past hour chasing ghosts and it seems to be
tied to something implicit around saving the config options after making
an update.  If someone sees the fatal error message, suggest that they
try saving their configuration, because that is what fixed things for
me after updating.