ci: use current tagged addon-builder
We were using `master` as the readme in the builder shows,
but at the time of writing, this has caused us to stop building armv7
slightly ahead of when the planned schedule for that was supposed to be.

Let's pin back to the current released version of the builder and
kick off a final build for armv7.

We can then revise this to follow releases in the intended way.
Maybe just leave it to dependabot?

refs: https://github.com/wez/govee2mqtt/issues/517
refs: https://github.com/home-assistant/builder/pull/263